Handle varargs method calls.
You may pass in either an array (as in the past) or individual Python arguments,
the match for a varargs method call is the minimum match for each of the
arguments.  Zero length arrays (i.e. no arguments) are also permitted with a
match value of 10.

Additionally, a jpy.type_translations dictionary is available with allows you to
easily wrap or modify all created objects of a particular java type.